requirements.txt
twilio>=9.0.0
python-dotenv>=1.0.0
requests>=2.31.0
click>=8.1.0
tqdm>=4.66.0
tabulate>=0.9.0
flask>=3.0.0
